requirements/production.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/production.txt requirements/production.in
#
attrs==19.3.0             # via packaging
certifi==2018.1.18        # via requests
chardet==3.0.4            # via requests
click==7.0                # via pip-tools
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
defusedxml==0.5.0         # via python3-openid, social-auth-core
diff-match-patch==20121119  # via django-import-export
django-extensions==2.2.1
django-import-export==1.2.0
django-social-authsch==0.1
django-solo==1.1.3
django==2.2.4
djangorestframework-api-key==1.4.1
djangorestframework==3.10.2
drf-yasg==1.16.1
entrypoints==0.3          # via flake8
et-xmlfile==1.0.1         # via openpyxl
flake8==3.7.8
future==0.18.2            # via python-language-server
gunicorn==19.7.1
idna==2.6                 # via requests
importlib-metadata==1.5.0  # via pluggy
inflection==0.3.1         # via drf-yasg
itypes==1.1.0             # via coreapi
jdcal==1.3                # via openpyxl
jedi==0.14.1              # via python-language-server
jinja2==2.11.1            # via coreschema
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
oauthlib==2.0.6           # via requests-oauthlib, social-auth-core
odfpy==1.3.6              # via tablib
openpyxl==2.5.0           # via tablib
packaging==19.1
parso==0.6.1              # via jedi
pillow==7.0.0
pip-tools==4.1.0
pluggy==0.13.1            # via python-language-server
psycopg2-binary==2.8.4
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pyjwt==1.5.3              # via social-auth-core
pyparsing==2.4.6          # via packaging
python-jsonrpc-server==0.3.4  # via python-language-server
python-language-server==0.28.2
python3-openid==3.1.0     # via social-auth-core
pytz==2017.3              # via django
pyyaml==5.3               # via tablib
requests-oauthlib==0.8.0  # via social-auth-core
requests==2.22.0          # via coreapi, requests-oauthlib, social-auth-core
ruamel.yaml.clib==0.2.0   # via ruamel.yaml
ruamel.yaml==0.16.7       # via drf-yasg
six==1.11.0               # via django-extensions, drf-yasg, packaging, pip-tools, social-auth-app-django, social-auth-core
social-auth-app-django==2.1.0  # via django-social-authsch
social-auth-core==1.6.0   # via django-social-authsch, social-auth-app-django
sqlparse==0.3.0           # via django
tablib==0.12.1            # via django-import-export
ujson==1.35               # via python-jsonrpc-server
unicodecsv==0.14.1        # via tablib
uritemplate==3.0.1        # via coreapi, drf-yasg
urllib3==1.25.8           # via requests
xlrd==1.1.0               # via tablib
xlwt==1.3.0               # via tablib
zipp==2.1.0               # via importlib-metadata
